+++ BusinessLogic/Material.cs
@@ -1,0 +1,9 @@
+
+namespace BusinessLogic
+{
+    public enum Material
+    {
+        Cemento = 5,
+        Ladrillo = 10 
+    }
+}
